+++ truffle/CHANGELOG.md
@@ -42,9 +42,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `Node.lookupContextReference(Class)` and `Node.lookupLanguageReference(Class)` that allows for a more convenient lookup.
 * Deprecated `RootNode.getLanguage(Class)`, the new language references should be used instead.
 * Added `TruffleFile` aware file type detector
-    - Added [TruffleFile.FileTypeDetector SPI](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleFile.FileTypeDetector.html) to detect a file MIME type and a file encoding.
+    - Added [TruffleFile.FileTypeDetector SPI](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleFile.FileTypeDetector.html) to detect a file MIME type and a file encoding. A language registering `FileTypeDetector` has to support all the MIME types recognized by the registered detector.
     - Added [TruffleFile.getMimeType method](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleFile.html#getMimeType--) to obtain a `TruffleFile` MIME type.
-    - Added [TruffleFile.getEncoding method](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleFile.html#getEncoding--) to obtain a `TruffleFile` encoding.
     - Added a possibility to set an [encoding in SourceBuilder](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/source/Source.SourceBuilder.html#encoding-java.nio.charset.Charset-)
     - The [Source builders](https://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/source/Source.html) are sandboxed for files and file URLs.
     - Deprecated usage of NIO `FileTypeDetector` for MIME type detection.